cleaning up code
