+++ yeat/tests/test_just_yeat_it.py
@@ -81,11 +81,11 @@ def test_invalid_input_algorithm(capsys, tmp_path):
 def test_invalid_custom_coverage_negative(coverage_depth):
     arglist = ["-c", coverage_depth, data_file("paired.cfg")]
     with pytest.raises(ArgumentError, match=rf"{coverage_depth} is not a positive integer"):
-        args = get_parser(exit_on_error=False).parse_args(arglist)
+        get_parser(exit_on_error=False).parse_args(arglist)
 
 
 @pytest.mark.parametrize("coverage_depth", [("string"), ("3.14")])
 def test_invalid_custom_coverage_noninteger(coverage_depth):
     arglist = ["-c", coverage_depth, data_file("paired.cfg")]
     with pytest.raises(ArgumentError, match=rf"{coverage_depth} is not an integer"):
-        args = get_parser(exit_on_error=False).parse_args(arglist)
+        get_parser(exit_on_error=False).parse_args(arglist)
